chore!: use doocs::Server instead of legacy interface
BREAKING CHANGE: The new interface no longer allows to create the same type of
location with arbitrary location codes, hence we have to decide for a fixed
code now. Using a location code which is not 10 (as used by default) now
results in an exception.
