+++ src/DoocsAdapter.cc
@@ -3,11 +3,23 @@
 
 #include "DoocsAdapter.h"
 
+#include "CSAdapterEqFct.h"
 #include "DoocsUpdater.h"
+#include "getAllVariableNames.h"
+#include "PropertyDescription.h"
+#include "VariableMapper.h"
+
+#include <filesystem>
 
 namespace ChimeraTK {
 
+  /*******************************************************************************************************************/
+
   std::atomic<bool> DoocsAdapter::isInitialised(false);
+
+  static char const* XML_CONFIG_SUFFIX = "-DoocsVariableConfig.xml";
+
+  /*******************************************************************************************************************/
 
   DoocsAdapter::DoocsAdapter() {
     // Create the managers. We need both
@@ -20,13 +32,19 @@ namespace ChimeraTK {
     updater = boost::make_shared<DoocsUpdater>();
   }
 
+  /*******************************************************************************************************************/
+
   boost::shared_ptr<DevicePVManager> const& DoocsAdapter::getDevicePVManager() const {
     return _devicePVManager;
   }
 
+  /*******************************************************************************************************************/
+
   boost::shared_ptr<ControlSystemPVManager> const& DoocsAdapter::getControlSystemPVManager() const {
     return _controlSystemPVManager;
   }
+
+  /*******************************************************************************************************************/
 
   void DoocsAdapter::waitUntilInitialised() {
     int i = 0;
@@ -41,6 +59,8 @@ namespace ChimeraTK {
     }
   }
 
+  /*******************************************************************************************************************/
+
   bool DoocsAdapter::checkPrintDataLossWarning(size_t counter) {
     // print first time at counter == 10 to suppress the messages spamming at startup
     if(counter < 1) return false;
@@ -51,6 +71,8 @@ namespace ChimeraTK {
     if(counter < 1000000) return counter % 100000 == 0;
     return counter % 1000000 == 0; // if the rate is 10Hz, this is roughly once per day
   }
+
+  /*******************************************************************************************************************/
 
   void DoocsAdapter::before_auto_init() {
     // prevent concurrent execution. It is unclear whether DOOCS may call auto_init in parallel in some situations, so
@@ -72,5 +94,144 @@ namespace ChimeraTK {
     }
     doocsAdapter.writeableVariablesWithMultipleProperties.clear(); // save memory (information no longer needed)
   }
+
+  /*******************************************************************************************************************/
+
+  std::unique_ptr<doocs::Server> DoocsAdapter::createServer() {
+    auto server = std::make_unique<doocs::Server>(ChimeraTK::ApplicationBase::getInstance().getName().c_str());
+
+    server->set_init_prolog([&] { this->eq_init_prolog(); });
+    server->set_post_init_epilog([&] { this->post_init_epilog(); });
+    server->set_cancel_epilog([&] { this->eq_cancel(); });
+
+    server->register_location_class<ChimeraTK::CSAdapterEqFct>();
+
+    return server;
+  }
+
+  /*******************************************************************************************************************/
+
+  /* eq_init_prolog is called before the locations are created, i.e. before the
+   * first call to eq_create. We initialise the application, i.e. all process
+   * variables are created in this function. */
+  void DoocsAdapter::eq_init_prolog() {
+    // set the DOOCS server name to the application name
+    // Create static instances for all applications cores. They must not have
+    // overlapping process variable names ("location/protery" must be unique).
+    ChimeraTK::ApplicationBase::getInstance().setPVManager(doocsAdapter.getDevicePVManager());
+    ChimeraTK::ApplicationBase::getInstance().initialise();
+
+    // the variable manager can only be filled after we have the CS manager
+    auto pvNames = ChimeraTK::getAllVariableNames(doocsAdapter.getControlSystemPVManager());
+
+    auto xmlFileName = ChimeraTK::ApplicationBase::getInstance().getName() + XML_CONFIG_SUFFIX;
+
+    if(std::filesystem::exists(xmlFileName)) {
+      ChimeraTK::VariableMapper::getInstance().prepareOutput(xmlFileName, pvNames);
+    }
+    else {
+      std::cerr << "WARNING: No XML file for the Doocs variable config found. Trying direct import." << std::endl;
+      ChimeraTK::VariableMapper::getInstance().directImport(pvNames);
+    }
+
+    // prepare list of unmapped read variables and pass it to the Application for optimisation
+    for(auto& p : ChimeraTK::VariableMapper::getInstance().getUsedVariables()) {
+      auto it = pvNames.find(p);
+      if(it != pvNames.end()) {
+        pvNames.erase(it);
+      }
+    }
+    ChimeraTK::ApplicationBase::getInstance().optimiseUnmappedVariables(pvNames);
+
+    // prepare list of properties connected to the same writable PV
+    {
+      // create map of PV name to properties using the PV (only for writable PVs)
+      std::map<std::string, std::set<std::shared_ptr<ChimeraTK::PropertyDescription>>> reverseMapping;
+      for(const auto& descr : ChimeraTK::VariableMapper::getInstance().getAllProperties()) {
+        for(const auto& source : descr->getSources()) {
+          if(!doocsAdapter.getControlSystemPVManager()->getProcessVariable(source)->isWriteable()) {
+            // to not add PVs to the mapping which are not writeable
+            continue;
+          }
+          reverseMapping[source].insert(descr);
+        }
+      }
+      // filter the map to contain only PVs being used at least twice with at least one writable property
+      for(const auto& p : reverseMapping) {
+        if(p.second.size() < 2) {
+          continue;
+        }
+        size_t writeableCount = 0;
+        for(const auto& d : p.second) {
+          auto attr = std::dynamic_pointer_cast<ChimeraTK::PropertyAttributes>(d);
+          if(attr->isWriteable) ++writeableCount;
+        }
+        if(writeableCount == 0) {
+          continue;
+        }
+        if(writeableCount > 1) {
+          // This case is not (yet) covered. It would require some synchronisation mechanism between writeable
+          // properties, which is not trivial as inconsistencies and dead locks must be avoided.
+          std::cout << "**** WARNING: Variable '" + p.first +
+                  "' mapped to more than one writeable property. Expect inconsistencies!\n";
+        }
+
+        // Add PVs which are used at least twice with at least one writable property to list
+        doocsAdapter.writeableVariablesWithMultipleProperties[p.first] = {};
+      }
+    }
+  }
+
+  /*******************************************************************************************************************/
+
+  /* post_init_epilog is called after all DOOCS properties are fully intialised,
+   * including any value intialisation from the config file. We start the
+   * application here. It will be launched in a separate thread. */
+  void DoocsAdapter::post_init_epilog() {
+    // check for locations not yet created (due to missing entries in the .conf file) and create them now
+    std::map<std::string, int> locMap = ChimeraTK::VariableMapper::getInstance().getLocationAndCode();
+    for(auto& loc : ChimeraTK::VariableMapper::getInstance().getAllLocations()) {
+      int codeToSet = 10;
+      bool defaultUsed = true;
+      // if no code is set in xml file, 10 is default
+      if(locMap.find(loc) != locMap.end()) {
+        codeToSet = locMap.find(loc)->second;
+        defaultUsed = false;
+      }
+
+      auto eq = find_device(loc);
+      if(eq == nullptr) {
+        add_location(codeToSet, loc);
+      }
+      else if(eq->fct_code() != codeToSet && !defaultUsed) {
+        throw ChimeraTK::logic_error("Location '" + loc + "' has already a fct_code '" +
+            std::to_string(eq->fct_code()) + "' from config file, that does not match with '" +
+            std::to_string(codeToSet) + "' from xml file");
+      }
+    }
+
+    // check for variables not yet initialised - we must guarantee that all to-application variables are written exactly
+    // once at server start.
+    for(auto& pv : doocsAdapter.getControlSystemPVManager()->getAllProcessVariables()) {
+      if(!pv->isWriteable()) continue;
+      if(pv->getVersionNumber() == ChimeraTK::VersionNumber(nullptr)) {
+        // The variable has not yet been written. Do it now, even if we just send a 0.
+        pv->write();
+      }
+    }
+
+    // start the application and the updater
+    ChimeraTK::ApplicationBase::getInstance().run();
+    doocsAdapter.updater->run();
+    ChimeraTK::DoocsAdapter::isInitialised = true;
+  }
+
+  /*******************************************************************************************************************/
+
+  void DoocsAdapter::eq_cancel() {
+    ChimeraTK::DoocsAdapter::isInitialised = false;
+  }
+
+  /*******************************************************************************************************************/
 
 } // namespace ChimeraTK
