--- src/empty_doocs_functions.cc
@@ -1,12 +1,0 @@
-// SPDX-FileCopyrightText: Deutsches Elektronen-Synchrotron DESY, MSK, ChimeraTK Project <chimeratk-support@desy.de>
-// SPDX-License-Identifier: LGPL-3.0-or-later
-
-// all the bloat we have to implement for DOOCS although we don't need it
-void eq_init_epilog() {}
-void refresh_prolog() {}
-void refresh_epilog() {}
-void interrupt_usr1_prolog(int) {}
-void interrupt_usr2_prolog() {}
-void interrupt_usr1_epilog(int) {}
-void interrupt_usr2_epilog() {}
-void post_init_prolog() {}
