+++ tests/serverTestCallOptimiseUnmappedVariables.conf
@@ -20,7 +20,7 @@ eq_fct_type:    10
 NAME:   "DOUBLE"
 }
 eq_fct_name:    "FLOAT"
-eq_fct_type:    11
+eq_fct_type:    10
 {
 NAME:   "FLOAT"
 }
