+++ tests/serverTestSpectrumBuffer.conf
@@ -13,6 +13,8 @@ eq_fct_type:    1
 SVR.RPC_NUMBER:         700000007
 SVR.NAME:       "PLAIN_VARIABLE_CREATION_TEST._SVR"
 SVR.BPN:        6000
+SVR.FACILITY: "TEST.DOOCS"
+SVR.DEVICE: "TEST"
 }
 eq_fct_name:    "INT"
 eq_fct_type:    10
