--- tests/serverTestVariableMapperWithLocationAndCode-DoocsVariableConfig.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<device_server xmlns="https://github.com/ChimeraTK/ControlSystemAdapter-DoocsAdapter">
-<!-- this is already in config file with code 10 -->
-  <location name="DOUBLE">
-    <D_spectrum source="TO_DEVICE_ARRAY"/>
-  </location>
-<!-- this is created new and shuold have code 10 (10 is default, if it is not set)-->
-  <location name="CREATED" >
-    <property name="COOL" source="/FLOAT/TO_DEVICE_SCALAR"/>
-  </location>
-<!-- this is created new but... -->
-  <location name="NEW" >
-    <property name="START" source="/FLOAT/TO_DEVICE_SCALAR"/>
-  </location>
-<!-- the code is set to 12 ( not default) -->
-  <location name="NEW" code="12" >
-    <property name="START2" source="/DOUBLE/TO_DEVICE_SCALAR"/>
-  </location>
-<!-- you can set the code, or not, but it wust be the same code as in config file -->
-  <location name="DOUBLE" code="10" >
-    <D_spectrum source="FROM_DEVICE_ARRAY">
-      <start>123.</start>
-      <increment>0.56</increment>
-    </D_spectrum>
-  </location>
- 
-  <location name="DOUBLE" code="10" >
-    <property name="POOL" source="/FLOAT/TO_DEVICE_SCALAR"/>
-  </location>
-</device_server>
