--- tests/serverTestVariableMapperWithLocationAndCode.conf
@@ -1,26 +1,0 @@
-eq_conf:
-
-oper_uid:       -1
-oper_gid:       405
-xpert_uid:      1000
-xpert_gid:      1000
-ring_buffer:    10000
-memory_buffer:  500
-
-eq_fct_name:    "OVERWRITE_CODE_TEST._SVR"
-eq_fct_type:    1
-{
-SVR.RPC_NUMBER:         700000102
-SVR.NAME:       "OVERWRITE_CODE_TEST._SVR"
-SVR.BPN:        6000
-}
-eq_fct_name:    "DOUBLE"
-eq_fct_type:    10
-{
-NAME:   "DOUBLE"
-}
-eq_fct_name:    "FLOAT"
-eq_fct_type:    11
-{
-NAME:   "FLOAT"
-}
