+++ tests/src/serverTestArrayPersistence.cpp
@@ -71,7 +71,8 @@ struct GlobalFixture {
   static std::string rpcNo;
   static std::string bpn;
   ThreadedDoocsServer server{boost::unit_test::framework::master_test_suite().p_name.value + ".conf",
-      boost::unit_test::framework::master_test_suite().argc, boost::unit_test::framework::master_test_suite().argv};
+      boost::unit_test::framework::master_test_suite().argc, boost::unit_test::framework::master_test_suite().argv,
+      doocsAdapter.createServer()};
 };
 
 ReferenceTestApplication GlobalFixture::referenceTestApplication{BOOST_STRINGIZE(BOOST_TEST_MODULE), alen};
